requirements.txt
google-adk[a2a,vertexai]
mcp
aiosqlite
python-dotenv
uvicorn
fastapi
google-cloud-aiplatform
